chore: move version related methods to versions file
