+++ internal/versions.go
@@ -420,7 +420,7 @@ func fetchBeatsBinary(ctx context.Context, artifactName string, artifact string,
 
 		maxTimeout := time.Duration(timeoutFactor) * time.Minute
 
-		downloadURL, err = utils.GetObjectURLFromBucket(bucket, prefix, object, maxTimeout)
+		downloadURL, err = getObjectURLFromBucket(bucket, prefix, object, maxTimeout)
 		if err != nil {
 			return "", err
 		}
@@ -434,6 +434,16 @@ func fetchBeatsBinary(ctx context.Context, artifactName string, artifact string,
 	}
 
 	return handleDownload(downloadURL)
+}
+
+func getBucketSearchNextPageParam(jsonParsed *gabs.Container) string {
+	token := jsonParsed.Path("nextPageToken")
+	if token == nil {
+		return ""
+	}
+
+	nextPageToken := token.Data().(string)
+	return "&pageToken=" + nextPageToken
 }
 
 // getGCPBucketCoordinates it calculates the bucket path in GCP
@@ -459,4 +469,139 @@ func getGCPBucketCoordinates(fileName string, artifact string) (string, string, 
 	}
 
 	return bucket, prefix, object
+}
+
+// getObjectURLFromBucket extracts the media URL for the desired artifact from the
+// Google Cloud Storage bucket used by the CI to push snapshots
+func getObjectURLFromBucket(bucket string, prefix string, object string, maxtimeout time.Duration) (string, error) {
+	exp := utils.GetExponentialBackOff(maxtimeout)
+
+	retryCount := 1
+
+	currentPage := 0
+	pageTokenQueryParam := ""
+	mediaLink := ""
+
+	storageAPI := func() error {
+		r := curl.HTTPRequest{
+			URL: fmt.Sprintf("https://storage.googleapis.com/storage/v1/b/%s/o?prefix=%s%s", bucket, prefix, pageTokenQueryParam),
+		}
+
+		response, err := curl.Get(r)
+		if err != nil {
+			log.WithFields(log.Fields{
+				"bucket":      bucket,
+				"elapsedTime": exp.GetElapsedTime(),
+				"prefix":      prefix,
+				"error":       err,
+				"object":      object,
+				"retry":       retryCount,
+			}).Warn("Google Cloud Storage API is not available yet")
+
+			retryCount++
+
+			return err
+		}
+
+		log.WithFields(log.Fields{
+			"bucket":      bucket,
+			"elapsedTime": exp.GetElapsedTime(),
+			"prefix":      prefix,
+			"object":      object,
+			"retries":     retryCount,
+			"url":         r.URL,
+		}).Trace("Google Cloud Storage API is available")
+
+		jsonParsed, err := gabs.ParseJSON([]byte(response))
+		if err != nil {
+			log.WithFields(log.Fields{
+				"bucket": bucket,
+				"prefix": prefix,
+				"object": object,
+			}).Warn("Could not parse the response body for the object")
+
+			retryCount++
+
+			return err
+		}
+
+		mediaLink, err = processBucketSearchPage(jsonParsed, currentPage, bucket, prefix, object)
+		if err != nil {
+			log.WithFields(log.Fields{
+				"currentPage": currentPage,
+				"bucket":      bucket,
+				"prefix":      prefix,
+				"object":      object,
+			}).Warn(err.Error())
+		} else if mediaLink != "" {
+			log.WithFields(log.Fields{
+				"bucket":      bucket,
+				"elapsedTime": exp.GetElapsedTime(),
+				"prefix":      prefix,
+				"medialink":   mediaLink,
+				"object":      object,
+				"retries":     retryCount,
+			}).Debug("Media link found for the object")
+			return nil
+		}
+
+		pageTokenQueryParam = getBucketSearchNextPageParam(jsonParsed)
+		if pageTokenQueryParam == "" {
+			log.WithFields(log.Fields{
+				"currentPage": currentPage,
+				"bucket":      bucket,
+				"prefix":      prefix,
+				"object":      object,
+			}).Warn("Reached the end of the pages and the object was not found")
+
+			return nil
+		}
+
+		currentPage++
+
+		log.WithFields(log.Fields{
+			"currentPage": currentPage,
+			"bucket":      bucket,
+			"elapsedTime": exp.GetElapsedTime(),
+			"prefix":      prefix,
+			"object":      object,
+			"retries":     retryCount,
+		}).Warn("Object not found in current page. Continuing")
+
+		return fmt.Errorf("The %s object could not be found in the current page (%d) the %s bucket and %s prefix", object, currentPage, bucket, prefix)
+	}
+
+	err := backoff.Retry(storageAPI, exp)
+	if err != nil {
+		return "", err
+	}
+	if mediaLink == "" {
+		return "", fmt.Errorf("Reached the end of the pages and the %s object was not found for the %s bucket and %s prefix", object, bucket, prefix)
+	}
+
+	return mediaLink, nil
+}
+
+func processBucketSearchPage(jsonParsed *gabs.Container, currentPage int, bucket string, prefix string, object string) (string, error) {
+	items := jsonParsed.Path("items").Children()
+
+	log.WithFields(log.Fields{
+		"bucket":  bucket,
+		"prefix":  prefix,
+		"objects": len(items),
+		"object":  object,
+	}).Debug("Objects found")
+
+	for _, item := range items {
+		itemID := item.Path("id").Data().(string)
+		objectPath := bucket + "/" + prefix + "/" + object + "/"
+		if strings.HasPrefix(itemID, objectPath) {
+			mediaLink := item.Path("mediaLink").Data().(string)
+
+			log.Infof("medialink: %s", mediaLink)
+			return mediaLink, nil
+		}
+	}
+
+	return "", fmt.Errorf("The %s object could not be found in the current page (%d) in the %s bucket and %s prefix", object, currentPage, bucket, prefix)
 }
